% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_polynom_phi.R
\name{Phi.v}
\alias{Phi.v}
\title{Calculation of the polynomials appearing in the CDF}
\usage{
Phi.v(v, x)
}
\arguments{
\item{v}{An integer >= 2.}

\item{x}{A real number between 0 and 1.}
}
\description{
Calculation of the polynomials appearing in the CDF
}
