% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probing_main.R
\name{probing}
\alias{probing}
\title{Title}
\usage{
probing(y, X, r)
}
\arguments{
\item{r}{}
}
\description{
Title
}
\examples{
theta <- c(runif(5,-5,5),rep(0,5))
w <- rnorm(15,0,0.02)
X <- replicate(10,rnorm(15,runif(1,-1,1),runif(1,0.05,0.3)))
y <- X\%*\%theta + w
X <- as.data.frame(X)
probing(y,X,r=0.1)
}
